--- .idea/copyright/profiles_settings.xml
@@ -1,7 +1,0 @@
-<component name="CopyrightManager">
-  <settings default="">
-    <LanguageOptions name="JAVA">
-      <option name="fileTypeOverride" value="1" />
-    </LanguageOptions>
-  </settings>
-</component>
