+++ app/build.gradle
@@ -43,4 +43,5 @@ dependencies {
     compile 'com.android.support:design:23.1.0'
     compile 'com.android.support:appcompat-v7:23.1.0'
     compile 'com.android.support:support-v13:23.1.0'
+    compile 'com.jakewharton:butterknife:7.0.1'
 }
