+++ app/src/androidTest/java/com/example/avjindersinghsekhon/minimaltodo/TestStoreRetrieveData.java
@@ -26,6 +26,10 @@ package com.example.avjindersinghsekhon.minimaltodo;
 import android.content.Context;
 import android.test.ActivityUnitTestCase;
 
+import com.example.avjindersinghsekhon.minimaltodo.activity.MainActivity;
+import com.example.avjindersinghsekhon.minimaltodo.business.StoreRetrieveData;
+import com.example.avjindersinghsekhon.minimaltodo.model.ToDoItem;
+
 import org.json.JSONArray;
 
 import java.util.ArrayList;
