+++ app/src/main/java/com/example/avjindersinghsekhon/minimaltodo/ItemTouchHelperClass.java
@@ -43,29 +43,4 @@ public class ItemTouchHelperClass extends ItemTouchHelper.Callback{
         adapter.onItemRemoved(viewHolder.getAdapterPosition());
 
     }
-
-//    @SuppressWarnings("deprecation")
-//    @Override
-//    public void onChildDraw(Canvas c, RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, boolean isCurrentlyActive) {
-//        if(actionState == ItemTouchHelper.ACTION_STATE_SWIPE){
-//            View itemView = viewHolder.itemView;
-//            Paint p = new Paint();
-//
-//
-//            MainActivity.BasicListAdapter.ViewHolder vh= (MainActivity.BasicListAdapter.ViewHolder)viewHolder;
-//            p.setColor(recyclerView.getResources().getColor(R.color.primary_light));
-//
-//            if(dX > 0){
-//                c.drawRect((float)itemView.getLeft(), (float)itemView.getTop(), dX, (float)itemView.getBottom(), p);
-//                String toWrite = "Left"+itemView.getLeft()+" Top "+itemView.getTop()+" Right "+dX+" Bottom "+itemView.getBottom();
-////                Log.d("OskarSchindler", toWrite);
-//            }
-//            else{
-//                String toWrite = "Left"+(itemView.getLeft()+dX)+" Top "+itemView.getTop()+" Right "+dX+" Bottom "+itemView.getBottom();
-////                Log.d("OskarSchindler", toWrite);
-//                c.drawRect((float) itemView.getRight() + dX, (float) itemView.getTop(), (float) itemView.getRight(), (float) itemView.getBottom(), p);
-//            }
-//            super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
-//        }
-//    }
 }
