--- app/src/main/java/com/example/avjindersinghsekhon/minimaltodo/SettingsActivity.java
@@ -1,66 +1,0 @@
-package com.example.avjindersinghsekhon.minimaltodo;
-
-import android.app.FragmentManager;
-import android.graphics.Color;
-import android.graphics.PorterDuff;
-import android.graphics.drawable.Drawable;
-import android.os.Bundle;
-import android.support.v4.app.NavUtils;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.MenuItem;
-
-public class SettingsActivity extends AppCompatActivity{
-
-    AnalyticsApplication app;
-    @Override
-    protected void onResume() {
-        super.onResume();
-        app.send(this);
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-
-        app = (AnalyticsApplication)getApplication();
-        String theme = getSharedPreferences(MainActivity.THEME_PREFERENCES, MODE_PRIVATE).getString(MainActivity.THEME_SAVED, MainActivity.LIGHTTHEME);
-        if(theme.equals(MainActivity.LIGHTTHEME)){
-            setTheme(R.style.CustomStyle_LightTheme);
-        }
-        else{
-            setTheme(R.style.CustomStyle_DarkTheme);
-        }
-        super.onCreate(savedInstanceState);
-//        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        setContentView(R.layout.activity_settings);
-        Toolbar toolbar = (Toolbar)findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        final Drawable backArrow = getResources().getDrawable(R.drawable.abc_ic_ab_back_mtrl_am_alpha);
-        if(backArrow!=null){
-            backArrow.setColorFilter(Color.WHITE, PorterDuff.Mode.SRC_ATOP);
-
-        }
-
-        if(getSupportActionBar()!=null){
-            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-            getSupportActionBar().setHomeAsUpIndicator(backArrow);
-        }
-
-        FragmentManager fm= getFragmentManager();
-        fm.beginTransaction().replace(R.id.mycontent, new SettingsFragment()).commit();
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch(item.getItemId()){
-            case android.R.id.home:
-                if(NavUtils.getParentActivityName(this)!=null){
-                    NavUtils.navigateUpFromSameTask(this);
-                }
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-}
