+++ app/src/main/java/com/example/avjindersinghsekhon/minimaltodo/service/TodoNotificationService.java
@@ -1,4 +1,4 @@
-package com.example.avjindersinghsekhon.minimaltodo;
+package com.example.avjindersinghsekhon.minimaltodo.service;
 
 import android.app.IntentService;
 import android.app.Notification;
@@ -7,6 +7,9 @@ import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 import android.util.Log;
+
+import com.example.avjindersinghsekhon.minimaltodo.R;
+import com.example.avjindersinghsekhon.minimaltodo.activity.ReminderActivity;
 
 import java.util.UUID;
