+++ app/src/main/java/com/example/avjindersinghsekhon/minimaltodo/view/CustomTextInputLayout.java
@@ -1,4 +1,4 @@
-package com.example.avjindersinghsekhon.minimaltodo;
+package com.example.avjindersinghsekhon.minimaltodo.view;
 
 
 import android.content.Context;
