+++ app/src/main/java/com/example/avjindersinghsekhon/minimaltodo/view/RecyclerViewEmptySupport.java
@@ -1,4 +1,4 @@
-package com.example.avjindersinghsekhon.minimaltodo;
+package com.example.avjindersinghsekhon.minimaltodo.view;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
