+++ app/src/main/res/layout/activity_main.xml
@@ -52,7 +52,7 @@
         <!--</android.support.design.widget.AppBarLayout>-->
 
 
-            <com.example.avjindersinghsekhon.minimaltodo.RecyclerViewEmptySupport
+            <com.example.avjindersinghsekhon.minimaltodo.view.RecyclerViewEmptySupport
                 app:layout_behavior="@string/appbar_scrolling_view_behavior"
                 android:id="@+id/toDoRecyclerView"
                 android:layout_width="match_parent"
