+++ app/src/main/res/layout/activity_todo_test.xml
@@ -28,7 +28,7 @@
                 android:layout_height="0dp">
 
 
-                <com.example.avjindersinghsekhon.minimaltodo.CustomTextInputLayout
+                <com.example.avjindersinghsekhon.minimaltodo.view.CustomTextInputLayout
                     app:hintTextAppearance="@style/FloatingLabelTextAppearance"
                     android:id="@+id/toDoCustomTextInput"
                     android:gravity="center"
@@ -48,7 +48,7 @@
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content" />
 
-                </com.example.avjindersinghsekhon.minimaltodo.CustomTextInputLayout>
+                </com.example.avjindersinghsekhon.minimaltodo.view.CustomTextInputLayout>
 
             </LinearLayout>
             <LinearLayout
